implement auth logic in controllers and secure routes

+++ backend/controllers/user.controller.js
@@ -1,5 +1,6 @@
 const db = require("../models");
 const bcrypt = require("bcryptjs");
+const jwt = require("jsonwebtoken");
 const userObject = db.users;
 
 exports.login = (req, res) => {
@@ -23,12 +24,19 @@ exports.login = (req, res) => {
         });
       }
 
+      const token = jwt.sign(
+        { id: user.id, role: user.role, email: user.email },
+        process.env.JWT_SECRET,
+        { expiresIn: "24h" }
+      );
+
       res.status(200).send({
         id: user.id,
         name: user.name,
         surname: user.surname,
         email: user.email,
-        role: user.role
+        role: user.role,
+        token: token
       });
     })
     .catch(err => {
@@ -76,9 +84,16 @@ exports.findAll = (req, res) => {
 exports.findOne = (req, res) => {
   const id = req.params.id;
 
+  if (req.user.id !== id && req.user.role !== "admin") {
+    return res.status(403).send({ message: "You can only view your own profile." });
+  }
+
   userObject
     .findOne({ where: { id: id } })
     .then((data) => {
+      if (!data) {
+        return res.status(404).send({ message: "User not found." });
+      }
       res.send(data);
     })
     .catch((err) => {
@@ -94,6 +109,10 @@ exports.update = (req, res) => {
   const surname = req.body.surname;
   const email = req.body.email;
   const password = req.body.password;
+
+  if (req.user.id !== id && req.user.role !== "admin") {
+    return res.status(403).send({ message: "You can only update your own profile." });
+  }
 
   const user = {
     name: name,
@@ -119,6 +138,10 @@ exports.update = (req, res) => {
 
 exports.delete = (req, res) => {
   const id = req.params.id;
+
+  if (req.user.id !== id && req.user.role !== "admin") {
+    return res.status(403).send({ message: "You can only delete your own profile." });
+  }
 
   userObject
     .destroy({ where: { id: id } })
